backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
websockets==12.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.7.0
tcgdex-sdk==2.2.1
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
boto3==1.35.99
anthropic==0.40.0
Pillow==11.3.0
imagehash==4.3.2
python-multipart==0.0.20
google-cloud-vision==3.13.0
google-auth==2.28.0
rapidfuzz==3.6.1
protobuf==6.33.6
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
geopy==2.4.1
tqdm==4.67.1
